@@ -1,17 +1,30 @@
 {
   "main": "node_modules/expo/AppEntry.js",
-  "private": true,
   "scripts": {
-    "start": "npm run debugger && exp start --lan --dev -i",
-    "debugger": "open 'rndebugger://set-debugger-loc?host=localhost&port=19001'"
+    "start": "expo start",
+    "android": "expo start --android",
+    "ios": "expo start --ios",
+    "web": "expo start --web",
+    "eject": "expo eject"
   },
   "dependencies": {
-    "@rematch/core": "^0.5.2",
-    "expo": "^25.0.0",
-    "prop-types": "^15.6.1",
-    "react": "16.2.0",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-25.0.0.tar.gz",
-    "react-native-status-bar-height": "^2.0.0",
-    "react-redux": "^5.0.7"
-  }
+    "@rematch/core": "^1.4.0",
+    "expo": "~36.0.0",
+    "expo-barcode-scanner": "~8.0.0",
+    "expo-constants": "~8.0.0",
+    "expo-font": "~8.0.0",
+    "expo-permissions": "~8.0.0",
+    "react": "~16.9.0",
+    "react-dom": "~16.9.0",
+    "react-native": "https://github.com/expo/react-native/archive/sdk-36.0.0.tar.gz",
+    "react-native-gesture-handler": "~1.5.0",
+    "react-native-status-bar-height": "^2.4.0",
+    "react-native-web": "~0.11.7",
+    "react-redux": "^7.2.0"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.0.0",
+    "babel-preset-expo": "~8.0.0"
+  },
+  "private": true
 }
